requirements.txt
Flask==3.0.2
requests==2.31.0
pandas==2.2.1
openpyxl==3.1.2
psutil==5.9.8
dnspython==2.4.2
python-whois==0.8.0
cryptography==42.0.2
pyOpenSSL==24.0.0
scapy==2.5.0 